moved selectTextView
From: removeContatByKey - this method is called when you call
removeContact:id, which is called usually from a tap on a
tableviewcell. this should not cause the keyboard to display

To: removeContactView - this method is only called internally when
contacts are removed via the backspace button. this is a situation
where you want the keyboard to stay up after the deletion

also note there was a second bug fix in this commit, namely if you
deleted the selectedContactView, it would leave
self.selectedContactView selected, which is used as a condition in
other places and then causes a mis-keyboard fire.